+++ src/scripts/run_tests.sh
@@ -2,6 +2,8 @@
 
 set -e
 
+export WORKSPACE_URL=https://ci.kbase.us/services/ws
+
 flake8 /app
 mypy --ignore-missing-imports /app
 bandit -r /app
