+++ src/test/mock_workspace/get_objects_41347.json
@@ -1,0 +1,35 @@
+{
+  "methods": ["POST"],
+  "path": "/",
+  "headers": {"Authorization": "valid_admin_token"},
+  "body": {
+    "version": "1.1",
+    "method": "Workspace.administer",
+    "params": [{
+      "command": "getObjects",
+      "params": {"objects": [{"ref": "41347/1/1"}], "no_data": 1}
+    }]
+  },
+  "response": {
+    "status": "200",
+    "body": {
+      "version": "1.1",
+      "result": [{
+        "data": [{
+          "info": [1, "Narrative.1553621013004", "KBaseNarrative.Narrative-4.0", "2019-03-26T17:23:33+0000", 1, "username", 41347, "username:narrative_1553621013004", "f4c8b4e9086d70dd24a8289f19a27914", 2121, {"creator": "username", "data_dependencies": "[]", "job_ids": "", "is_temporary": "true", "format": "ipynb", "name": "Untitled", "description": "", "type": "KBaseNarrative.Narrative", "ws_name": "username:narrative_1553621013004"}],
+
+          "path": ["41347/1/1"],
+          "provenance": [{"script": "NarrativeManager.py", "input_ws_objects": [], "resolved_ws_objects": [], "external_data": [], "subactions": [], "custom": {}, "description": "Created new Workspace/Narrative bundle."}],
+          "creator": "username",
+          "orig_wsid": 41347,
+          "created": "2019-03-26T17:23:33+0000",
+          "epoch": 1553621013061,
+          "refs": [],
+          "copy_source_inaccessible": 0,
+          "extracted_ids": {}
+        }]
+      }]
+    }
+  }
+}
+
