+++ src/test/mock_workspace/get_permissions_mass_39794.json
@@ -1,0 +1,26 @@
+{
+  "methods": ["POST"],
+  "path": "/",
+  "headers": {"Authorization": "valid_admin_token"},
+  "body": {
+    "version": "1.1",
+    "method": "Workspace.administer",
+    "params": [{
+      "command": "getPermissionsMass",
+      "params": {"workspaces": [{"id": 39794}]}
+    }]
+  },
+  "response": {
+    "status": "200",
+    "body": {
+      "version": "1.1",
+      "result": [{
+        "perms": [
+          {"username": "a"},
+          {"username2": "w"},
+          {"*": "r"}
+        ]
+      }]
+    }
+  }
+}
