+++ src/test/mock_workspace/get_permissions_mass_41348.json
@@ -1,0 +1,22 @@
+{
+  "methods": ["POST"],
+  "path": "/",
+  "headers": {"Authorization": "valid_admin_token"},
+  "body": {
+    "version": "1.1",
+    "method": "Workspace.administer",
+    "params": [{
+      "command": "getPermissionsMass",
+      "params": {"workspaces": [{"id": 41348}]}
+    }]
+  },
+  "response": {
+    "status": "200",
+    "body": {
+      "version": "1.1",
+      "result": [{
+        "perms": [{"username": "a"}]
+      }]
+    }
+  }
+}
