+++ src/test/mock_workspace/get_workspace_info_15.json
@@ -1,0 +1,22 @@
+{
+  "methods": ["POST"],
+  "path": "/",
+  "headers": {"Authorization": "valid_admin_token"},
+  "body": {
+    "version": "1.1",
+    "method": "Workspace.administer",
+    "params": [{
+      "command": "getWorkspaceInfo",
+      "params": {"id": 15}
+    }]
+  },
+  "response": {
+    "status": "200",
+    "body": {
+      "version": "1.1",
+      "result": [
+        [15, "username:narrative_1553621013004", "username", "2019-04-05T22:17:00+0000", 12, "n", "r", "unlocked", {"searchtags": "narrative", "is_temporary": "false", "narrative": "1", "data_palette_id": "2", "cell_count": "1", "narrative_nice_name": "wat"}]
+      ]
+    }
+  }
+}
