+++ src/test/test_data/reads_save_15_44.json
@@ -1,0 +1,103 @@
+{
+  "ws_info": {"version":"1.1","result":[[15,"KBaseExampleData","kbasetest","2016-10-05T17:16:20+0000",45,"n","r","unlocked",{}]]},
+  "obj": {
+    "data": {
+      "insert_size_mean": null,
+      "insert_size_std_dev": null,
+      "interleaved": 1,
+      "lib1": {
+        "encoding": "ascii",
+        "file": {
+          "file_name": "235066e6-35d8-49e5-a7da-b012934835a6.inter.fastq.gz",
+          "hid": "KBH_55378",
+          "id": "3dc199fa-50f8-4a04-92eb-a582b8cd999d",
+          "remote_md5": "9ebbe6b4bc334535bc4ab968d9b32e2a",
+          "type": "shock",
+          "url": "https://ci.kbase.us/services/shock-api"
+        },
+        "size": 36510129,
+        "type": "fq"
+      },
+      "read_orientation_outward": 0,
+      "sequencing_tech": "Illumina",
+      "single_genome": 1
+    },
+    "info": [
+      44,
+      "rhodobacter.art.q20.int.PE.reads",
+      "KBaseFile.PairedEndLibrary-2.0",
+      "2016-10-05T17:11:40+0000",
+                  1,
+                  "kbasetest",
+                  15,
+                  "KBaseExampleData",
+                  "abaa39d622962856b799c3eb32de2c2f",
+                  447,
+                  {
+                    "sequencing_tech": "Illumina",
+                    "single_genome": "1"
+                  }
+    ],
+    "path": [
+      "15/44/1"
+    ],
+    "provenance": [
+      {
+        "time": "2016-10-05T13:00:55+0000",
+        "epoch": 1475672455000,
+        "service": "legacy_reads_conversion",
+        "service_ver": "0.0.1",
+        "method": "run_legacy_reads_conversion",
+        "method_params": [
+          {
+            "output": "rhodobacter.art.q20.int.PE.reads",
+            "workspace_name": "pranjan77:1475663060070",
+            "single_genome": 1,
+            "sequencing_tech": "Illumina",
+            "read_library_name": "rhodo.art.q20.int.PE.reads"
+          }
+        ],
+        "input_ws_objects": [
+          "11802/2/7"
+        ],
+        "resolved_ws_objects": [
+          "11802/2/7"
+        ],
+        "external_data": [],
+        "subactions": [
+          {
+            "name": "legacy_reads_conversion",
+            "ver": "0.0.1",
+            "code_url": "https://github.com/pranjan77/legacy_reads_conversion",
+            "commit": "ab6819d9d49a8866f52b216df1d97bd42a56a7aa"
+          },
+          {
+            "name": "ReadsUtils",
+            "ver": "0.1.2-dev",
+            "code_url": "https://github.com/kbaseapps/ReadsUtils",
+            "commit": "98ce5075c2117cec62505200ace121bc736d9569"
+          },
+          {
+            "name": "DataFileUtil",
+            "ver": "0.0.6-release",
+            "code_url": "https://github.com/kbaseapps/DataFileUtil",
+            "commit": "b850c6cca09f85e9f31afe42b389b96fa1b8a42a"
+          }
+        ],
+        "custom": {},
+        "description": "KBase SDK method run via the KBase Execution Engine"
+      }
+    ],
+    "creator": "pranjan77",
+    "orig_wsid": 11802,
+    "created": "2016-10-05T13:04:11+0000",
+    "epoch": 1475672651086,
+    "refs": [],
+    "copy_source_inaccessible": 1,
+    "extracted_ids": {
+      "handle": [
+        "KBH_55378"
+      ]
+    }
+  }
+}
